pyproject.toml
[project]
name = "d-schema"
version = "0.1.0"
description = "A Python library to generate database schemas in various formats."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasketch>=1.5.9",
    "mysql-connector-python>=9.4.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.0",
    "hydra-core>=1.3.2",
    "importlib-resources>=6.0.0; python_version < '3.9'",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"d_schema" = ["config/**/*.yaml"]
